+++ demo-blogas/src/main/java/com/example/demoblogas/comments/CommentsController.java
@@ -44,7 +44,7 @@ public class CommentsController {
 
 
     @PostMapping("/blogs/{id}/comment")
-    public String createNewComment(@PathVariable("id") int id, @Valid Comment comment, BindingResult errors) {
+    public String createNewComment(@PathVariable("id") Long id, @Valid Comment comment, BindingResult errors) {
         if (errors.hasErrors()) {
             return "comments/comment";
         }
